fetcher/requirements.txt
aioquic==1.2.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
fastapi==0.115.13
Flask==3.1.0
h11==0.16.0
h2==4.1.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
kombu==5.5.4
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==12.1.1
mitmproxy-macos==0.12.6
mitmproxy_rs==0.12.6
msgpack==1.1.0
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.51
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pylsqpack==0.3.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
tornado==6.5
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
urwid==2.6.16
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
